+++ cli/cli_test.go
@@ -17,27 +17,17 @@ package cli
 
 import (
 	"bytes"
-	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
 	"testing"
-	"time"
 
-	posmoni "github.com/NethermindEth/posmoni/pkg/eth2"
-	"github.com/NethermindEth/sedge/configs"
 	"github.com/NethermindEth/sedge/internal/pkg/commands"
-	"github.com/NethermindEth/sedge/internal/utils"
 	"github.com/NethermindEth/sedge/test"
 	"github.com/NethermindEth/sedge/test/mock_prompts"
 	"github.com/golang/mock/gomock"
 	log "github.com/sirupsen/logrus"
-)
-
-var (
-	inspectExecutionUrl = "http://192.168.128.3"
-	inspectConsensusUrl = "http://192.168.128.3"
 )
 
 var inspectOut = `
@@ -116,7 +106,6 @@ type cliCmdTestCase struct {
 	name       string
 	configPath string
 	runner     commands.CommandRunner
-	monitor    MonitoringTool
 	fdOut      *bytes.Buffer
 	args       CliCmdFlags
 	isPreErr   bool
@@ -172,9 +161,6 @@ func prepareCliCmd(tc cliCmdTestCase) {
 	// Set config file path
 	cfgFile = tc.configPath
 	initConfig()
-	initMonitor(func() MonitoringTool {
-		return tc.monitor
-	})
 }
 
 func buildCliTestCase(
@@ -211,25 +197,6 @@ func buildCliTestCase(
 		},
 		SRunBash: func(bs commands.ScriptFile) (string, error) {
 			return "", nil
-		},
-	}
-
-	// Check for port occupation
-	defaultsPorts := map[string]string{
-		"ELApi": configs.DefaultApiPortEL,
-		"CLApi": configs.DefaultApiPortCL,
-	}
-	ports, err := utils.AssignPorts("localhost", defaultsPorts)
-	if err != nil {
-		t.Fatalf(configs.PortOccupationError, err)
-	}
-
-	tc.monitor = &monitorStub{
-		data: []posmoni.EndpointSyncStatus{
-			{Endpoint: inspectExecutionUrl + ":" + ports["ELApi"], Synced: true},
-			{Endpoint: inspectConsensusUrl + ":" + ports["CLApi"], Synced: true},
-			{Endpoint: inspectExecutionUrl + ":" + ports["ELApi"], Synced: true},
-			{Endpoint: inspectConsensusUrl + ":" + ports["CLApi"], Synced: true},
 		},
 	}
 
@@ -411,180 +378,6 @@ func TestCliCmd(t *testing.T) {
 			} else if tc.isErr && err == nil {
 				t.Errorf("%s expected to fail", descr)
 			}
-		})
-	}
-}
-
-// Stub for MonitoringTool interface
-type monitorStub struct {
-	data  []posmoni.EndpointSyncStatus
-	calls int
-}
-
-func (ms *monitorStub) TrackSync(done <-chan struct{}, beaconEndpoints, executionEndpoints []string, wait time.Duration) <-chan posmoni.EndpointSyncStatus {
-	ms.calls++
-	c := make(chan posmoni.EndpointSyncStatus, len(ms.data))
-	var w time.Duration
-
-	go func() {
-		for {
-			select {
-			case <-done:
-				close(c)
-				return
-			case <-time.After(w):
-				if w == 0 {
-					// Don't wait the first time
-					w = wait
-				}
-				for _, d := range ms.data {
-					c <- d
-				}
-			}
-		}
-	}()
-
-	return c
-}
-
-func TestTrackSync(t *testing.T) {
-	t.Parallel()
-
-	tcs := []struct {
-		name    string
-		data    []posmoni.EndpointSyncStatus
-		runner  commands.CommandRunner
-		flags   CliCmdFlags
-		isError bool
-	}{
-		{
-			"Test case 1, execution client got an error",
-			[]posmoni.EndpointSyncStatus{
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: false, Error: errors.New("")},
-			},
-			&test.SimpleCMDRunner{},
-			CliCmdFlags{
-				generationPath: configs.DefaultDockerComposeScriptsPath,
-			},
-			true,
-		},
-		{
-			"Test case 2, execution client got an error, consensus client not synced",
-			[]posmoni.EndpointSyncStatus{
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: false, Error: errors.New("")},
-				{Endpoint: configs.OnPremiseConsensusURL, Synced: false},
-			},
-			&test.SimpleCMDRunner{},
-			CliCmdFlags{
-				generationPath: configs.DefaultDockerComposeScriptsPath,
-			},
-			true,
-		},
-		{
-			"Test case 3, execution client got an error, consensus client synced",
-			[]posmoni.EndpointSyncStatus{
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: false, Error: errors.New("")},
-				{Endpoint: configs.OnPremiseConsensusURL, Synced: true},
-			},
-			&test.SimpleCMDRunner{},
-			CliCmdFlags{
-				generationPath: configs.DefaultDockerComposeScriptsPath,
-			},
-			true,
-		},
-		{
-			"Test case 4, bad execution client response, good consensus client response",
-			[]posmoni.EndpointSyncStatus{
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: true, Error: errors.New("")},
-				{Endpoint: configs.OnPremiseConsensusURL, Synced: true},
-			},
-			&test.SimpleCMDRunner{},
-			CliCmdFlags{
-				generationPath: configs.DefaultDockerComposeScriptsPath,
-			},
-			true,
-		},
-		{
-			"Test case 5, consensus client got an error, consensus client not synced",
-			[]posmoni.EndpointSyncStatus{
-				{Endpoint: configs.OnPremiseConsensusURL, Synced: false, Error: errors.New("")},
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: false},
-			},
-			&test.SimpleCMDRunner{},
-			CliCmdFlags{
-				generationPath: configs.DefaultDockerComposeScriptsPath,
-			},
-			true,
-		},
-		{
-			"Test case 6, consensus client got an error, consensus client synced",
-			[]posmoni.EndpointSyncStatus{
-				{Endpoint: configs.OnPremiseConsensusURL, Synced: false, Error: errors.New("")},
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: true},
-			},
-			&test.SimpleCMDRunner{},
-			CliCmdFlags{
-				generationPath: configs.DefaultDockerComposeScriptsPath,
-			},
-			true,
-		},
-		{
-			"Test case 7, mixed results",
-			[]posmoni.EndpointSyncStatus{
-				{Endpoint: configs.OnPremiseConsensusURL, Synced: false},
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: true},
-				{Endpoint: configs.OnPremiseConsensusURL, Synced: false},
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: true},
-				{Endpoint: configs.OnPremiseConsensusURL, Synced: true},
-			},
-			&test.SimpleCMDRunner{},
-			CliCmdFlags{
-				generationPath: configs.DefaultDockerComposeScriptsPath,
-			},
-			false,
-		},
-		{
-			"Test case 8, mixed results",
-			[]posmoni.EndpointSyncStatus{
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: false},
-				{Endpoint: configs.OnPremiseConsensusURL, Synced: true},
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: false},
-				{Endpoint: configs.OnPremiseConsensusURL, Synced: true},
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: false},
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: false},
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: false},
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: true},
-			},
-			&test.SimpleCMDRunner{},
-			CliCmdFlags{
-				generationPath: configs.DefaultDockerComposeScriptsPath,
-			},
-			false,
-		},
-		{
-			"Test case 9, mixed results, error",
-			[]posmoni.EndpointSyncStatus{
-				{Endpoint: configs.OnPremiseConsensusURL, Synced: false},
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: true},
-				{Endpoint: configs.OnPremiseConsensusURL, Synced: false},
-				{Endpoint: configs.OnPremiseExecutionURL, Synced: true},
-				{Endpoint: configs.OnPremiseConsensusURL, Synced: false, Error: errors.New("")},
-			},
-			&test.SimpleCMDRunner{},
-			CliCmdFlags{
-				generationPath: configs.DefaultDockerComposeScriptsPath,
-			},
-			true,
-		},
-	}
-	// TODO: Starvation bc a synced status from one of the clients is not tested
-
-	for _, tc := range tcs {
-		t.Run(tc.name, func(t *testing.T) {
-			ms := monitorStub{data: tc.data}
-
-			err := trackSync(tc.runner, &ms, "", "", time.Millisecond*100, &tc.flags)
-			utils.CheckErr("trackSync(...) failed", tc.isError, err)
 		})
 	}
 }